"Updated migration to use unsignedBigInteger for user_id and poll_id, and changed header text in home.blade.php"

+++ database/migrations/2024_01_22_182113_create_user_poll_votes_table.php
@@ -14,8 +14,8 @@ return new class extends Migration {
     {
         Schema::create('user_poll_votes', function (Blueprint $table) {
             $table->id();
-            $table->Integer('user_id');
-            $table->Integer('poll_id');
+            $table->unsignedBigInteger('user_id');
+            $table->unsignedBigInteger('poll_id');
             $table->Integer('option_id')->nullable();
             $table->timestamps();
